接口更新

+++ server/api/manage.go
@@ -21,20 +21,30 @@ func Delete(w http.ResponseWriter, r *http.Request) {
 	r.ParseForm()
 	if r.Method == "POST" {
 		//err := json.Unmarshal([]byte(r.PostFormValue("ids")), &ids)
-		ids := r.Form.Get("ids")
-		if ids != "" {
-			ids = strings.ReplaceAll(ids, "[", "(")
-			ids = strings.ReplaceAll(ids, "]", ")")
-			engine := utils.GetXORMEngine()
-			defer engine.Close()
-			sql := fmt.Sprintf("DELETE FROM cili_engine WHERE id IN %s;", ids)
-			_, err := engine.Exec(sql)
-			if err == nil {
-				resp, _ := utils.Json(models.BaseResp{
-					Msg: "删除成功",
-				})
-				fmt.Fprintf(w, resp)
+		token := utils.DecryptHexAesStr(r.Form.Get("token"))
+		c, _ := utils.ParseToken(token)
+		if utils.GetTokenVal(c, "remove") == 1 {
+
+			ids := r.Form.Get("ids")
+			if ids != "" {
+				ids = strings.ReplaceAll(ids, "[", "(")
+				ids = strings.ReplaceAll(ids, "]", ")")
+				engine := utils.GetXORMEngine()
+				defer engine.Close()
+				sql := fmt.Sprintf("DELETE FROM cili_engine WHERE id IN %s;", ids)
+				_, err := engine.Exec(sql)
+				if err == nil {
+					resp := utils.JsonString(models.BaseResp{
+						Msg: "删除成功",
+					})
+					fmt.Fprintf(w, resp)
+				}
 			}
+		} else {
+			fmt.Fprintf(w, utils.JsonString(models.BaseResp{
+				Code:   403,
+				ErrMsg: "没有删除权限",
+			}))
 		}
 	}
 }
